pyproject.toml
[project]
name = "macrobania"
version = "0.0.1"
description = "Context-aware macro via Vision-Language Agents — record once, replay anywhere."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Apache-2.0" }
authors = [{ name = "macro-bania contributors" }]
keywords = ["vlm", "vla", "gui-agent", "macro", "automation", "rpa", "qwen"]
classifiers = [
    "Development Status :: 1 - Planning",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: Microsoft :: Windows :: Windows 10",
    "Operating System :: Microsoft :: Windows :: Windows 11",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # CLI
    "click>=8.1",
    "rich>=13.7",
    # 관측/로그
    "structlog>=24.1",
    # 설정
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "tomli; python_version<'3.11'",
    # HTTP / VLM 클라이언트 (OpenAI 호환 엔드포인트)
    "httpx>=0.27",
    "openai>=1.30",
    # 이미지
    "pillow>=10.3",
    "numpy>=1.26",
]

[project.optional-dependencies]
# Phase 1: Recorder
capture = [
    "dxcam>=0.0.5; sys_platform == 'win32'",
    "pynput>=1.7.6",
    "mss>=9.0",            # 폴백 캡처
    "pywinauto>=0.6.8; sys_platform == 'win32'",
    "uiautomation>=2.0.18; sys_platform == 'win32'",
    "psutil>=5.9",
]

# OCR
ocr = [
    "rapidocr-onnxruntime>=1.3",
    "onnxruntime>=1.17",
]

# 입력 주입 (Play)
inject = [
    "pydirectinput>=1.0.4; sys_platform == 'win32'",
    "pyautogui>=0.9.54",
]

# UI (Phase 5)
ui = [
    "PySide6>=6.6",
]

# 안전장치
safety = [
    "presidio-analyzer>=2.2",
    "presidio-anonymizer>=2.2",
]

# 개발/테스트
dev = [
    "pytest>=8.1",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "pytest-timeout>=2.3",
    "ruff>=0.4",
    "mypy>=1.10",
    "ipykernel>=6.29",
]

# 전체
all = [
    "macrobania[capture,ocr,inject,ui,safety,dev]",
]

[project.scripts]
macrobania = "macrobania.cli:main"

[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/macrobania"]

# --- Ruff ---
[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "SIM", "RUF"]
ignore = ["E501"]  # line-length는 formatter가 관리

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# --- Mypy ---
[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_ignores = true
warn_return_any = true
disallow_untyped_defs = true
files = ["src/macrobania"]

# --- Pytest ---
[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra --strict-markers --cov=macrobania --cov-report=term-missing --timeout=20"
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.coverage.run]
branch = true
source = ["macrobania"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
